auth-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
motor>=3.3.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
# Auth & security
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
python-multipart>=0.0.9
# Email
fastapi-mail>=1.4.0
aiosmtplib>=3.0.0
# Google OAuth
google-auth>=2.28.0
httpx>=0.27.0
requests>=2.31.0
# Utils
python-dotenv>=1.0.0
cloudinary>=1.39.0
apscheduler>=3.10.0
sqlalchemy>=2.0.0
